+++ CHANGES.md
@@ -15,6 +15,8 @@ option.
 * VIM-483 Added support for `number` option
 * VIM-624 Deselect visual selection range on opening the Ex entry field
 * VIM-511 Fixed editing offset after `<BS>` for `.` command
+* VIM-792 Fixed line-wise and block-wise paste commands for `*` and `+`
+  registers
 
 
 0.37, 2014-10-15